requirements.txt
# streamlit==1.29.0
# scikit-learn==1.3.2
# numpy==1.26.4
# python-docx==1.1.0
# PyPDF2==3.0.1


streamlit==1.28.0
scikit-learn==1.8.0
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.0.0
joblib==1.3.2
pandas==2.3.3
openpyxl==3.1.5
spacy==3.7.2
plotly==5.18.0
matplotlib>=3.7.0
pdf2image>=1.16.0
pytesseract>=0.3.10